+++ CesiumRasterOverlays/src/RasterOverlay.cpp
@@ -34,7 +34,7 @@ public:
             ellipsoid) {}
 
   virtual CesiumAsync::Future<LoadedRasterOverlayImage>
-  loadTileImage(RasterOverlayTile& /* overlayTile */) override {
+  loadTileImage(const RasterOverlayTile& /* overlayTile */) override {
     return this->getAsyncSystem()
         .createResolvedFuture<LoadedRasterOverlayImage>({});
   }